Add "Save as" dialog, and call it when we need a filename
Make LatexDocument.dir into a property.  When it's accessed, and it
isn't set, pop up a "Save as" dialog to get the directory and filename.
If the user doesn't choose a file, raise an ObstinateUserError.  Since
the fullfilename properties call dir, we can use the fullfilename and
know it is either set or an exception will be raised elsewhere.

In the future, we should probably make sure a filename is chosen on
loading, so we're not exposed to these problems.