Mark presentation as modified when slide order changes
Rather than do this for each event that changes the slide order, we hook
into the on_row_inserted/on_row_deleted methods.  But we don't want to
catch the rows being created during loading, so an attribute _loaded is
added and set to False during the loading process.  We only need to set
a single modification flag to True, not all of them.  We choose header
for this.  In get_modified, we allow ourselves to short-circuit if the
header is modified, since the pages may not be set up with LatexSlides
when the slide order has been changed.