Set up loading and saving of document settings
Nothing is done with the settings yet.  We use an OrderedDict to keep
the skeletons in order for later use.  The format of a SlideDeX file has
changed: The slide separator is now '%%SLIDEDEX%%'.  It appears at the
beginning of the file (followed by a currently non-conserved
description).  It is followed by a commented settings block and another
separator before the header.

Beginning of solution to bug #8.