Add Menu to insert slide with choice of skeleton
DocumentSettings calls LatexDocument.generate_skeleton_menu after
reading in the settings.  This creates a gtk.Menu populated with
MenuItems corresponding to the skeletons.  The first of these is hooked
up to the Shift+Insert accelerator.  This menu is connected to the
menubar and the toolbar button.

This mostly solves issue #7.