Compile individual slides that need it when compiling whole doc
This closes #14.  To do this, we need to keep track of two modification
states: modification since save and modification since compile.  This
necessates rewriting the entire modification tracking code, so I
switched over to using properties while I was at it.

A happy side effect of this is that the slides will be automatically
compiled on the first load, since we mark them as modified here.  This
means that we generally won't need to render newly added slides, so we
switch the default on this option to False.

The header and footer make for an odd corner case.  They aren't
compiled, but when they're changed, you'll want to recompile everything.
This is far enough from the standard slide behavior that they now have
their own class.