Add drag and drop handling
Two different mechanisms are enabled.  When moving pages within a
presentation, we just reorder them.  When copying pages within or
between presentations, we send the text of the pages across and create
new LatexSlides for them.  These are set to render right away.  This is
perhaps unnecessary when copying slides within a presentation, since we
could just reuse the existing pixbuf, but it's easier to leave it this
way.

Closes #4 and closes #12.  What a commit!