Find modifications since save by file modification time
When closing, delete all temp files that have been modified after
the main tex file.  It's likely that they have changes that haven't
been saved.

This can result in temp files not being saved when they are created
during load, but the whole presentation is never compiled.  But this
seems like a relatively rare case.